Add files via upload
The purpose of this contribution is to simulate brown motion moves and collisions and save it e.g. to gif file by pillowwriter tool.

1. Theory

Brownian motion - the random motion of a particle as a result of collisions with surrounding atoms or molecules. Diffusiophoresis is the movement of a group of particles induced by a concentration gradient. This movement always flows from areas of high concentration to areas of low concentration. Observed by Scottish botanist Robert Brown.

Euler Maruyama method - method for the approximate numerical solution of a stochastic differential equation (SDE). It is an extension of the Euler method for ordinary differential equations to stochastic differential equations. It is named after Leonhard Eulerand Gisiro Maruyama. Unfortunately, the same generalization cannot be done for any arbitrary deterministic method.

Geometric (exponential) brownian motion - stochastic process, is a continuous-time stochastic, the logarithm of the randomly varying quantity follows a Brownian motion.  

Perfect elasticity - the price is the same (constant), no matter what the quantity is. Perfectly inelastic demand is e.g. the demand for cheap essential goods, salt or water i safe locations. 

Wiener process -  the same as brown motion, but observed by American mathematician Norbert Wiener.



3. References

https://medium.com/@polanitzer/geometric-brownian-motion-in-python-predict-the-bitcoin-prices-6d7e34d9b407 
https://towardsdatascience.com/how-to-simulate-financial-portfolios-with-python-d0dc4b52a278 
https://towardsdatascience.com/how-to-simulate-a-stock-market-with-less-than-10-lines-of-python-code-5de9336114e5
https://stackoverflow.com/questions/45021301/geometric-brownian-motion-simulation-in-python
https://stackoverflow.com/questions/61569019/plotting-gbm-issues
https://ipython-books.github.io/133-simulating-a-brownian-motion/
https://en.wikipedia.org/wiki/Euler–Maruyama_method 
https://github.com/mCodingLLC/VideosSampleCode/tree/master/videos
https://www.youtube.com/watch?v=hqSnruUe3tA
https://www.codearmo.com/blog/pricing-options-monte-carlo-simulation-python
https://en.wikipedia.org/wiki/Brownian_motion 
https://towardsdatascience.com/animated-visualization-of-brownian-motion-in-python-3518ecf28533
https://www.sciencedirect.com/topics/chemistry/brownian-motion 
https://medium.com/analytics-vidhya/modelling-geometric-brownian-motion-in-python-b65462cc2c1d
https://github.com/neilyejjey1999/yejjeyprojects/blob/main/FIN5615_Project_2_Neil_Yejjey.ipynb
https://github.com/cantaro86/Financial-Models-Numerical-Methods
https://github.com/cantaro86/Financial-Models-Numerical-Methods/blob/master/1.1%20Black-Scholes%20numerical%20methods.ipynb
https://scipython.com/blog/two-dimensional-collisions/
https://medium.com/analytics-vidhya/building-a-monte-carlo-method-stock-price-simulator-with-geometric-brownian-motion-and-bootstrap-e346ff464894
https://en.wikipedia.org/wiki/Wiener_process